Rethinking library architecture and disposable resource management
- Before this commit, ONIXController held a IDisposable resource
(acquisition context) at build and edit time. This lead to a tension
between the indented design of the Bonsai language and the way
we were using it.
- The issue boils down to the fact that we want to use the editor to
manipulate hardware state before running the workflow because its very
convenient. This means that public properties of workflow elements are
often straight functions of hardware state and do not exist in
intermediate memory owned by the class.
- At the same time, we like all the benefits of public properties as
they are intended to be used during workflow operation, e.g. as optional
reactive sinks and sources.
- The issue is that disposable resources can only be properly handled
within a single function call scope with the `using` statement or at
runtime using reactive operators designed to deal with disposable
resources.
- So, in the old design, since the disposable resource was held by nodes
at the top level scope and those nodes can not implement IDisposable
since it will not be properly used by the Bonsai "runtime", we were
leaking resources.
- This commit deals with this by limiting the lifetime of disposable
resources to withing using statements or during runs of the workflow,
and nowhere else. It does this by sharing a key for accessing the
disposable resource (oni.Context) but not the context itself. The
downside is that during edit time, the context must be recreated for
each and every access to hardware. This is a major UI performance issue
and probably needs to be addressed. This is why this is on a separate
branch for the time being.
- I have a post on the Bonsai forum about this here:
https://groups.google.com/g/bonsai-users/c/8EGR9swedTI