Add ONIXDeviceDescriptor
- ONIXDeviceDescriptor is an attempt to make the library more robust to
hardware changes that can occur in between saving and opening workflows,
disconnected headstages, etc.
- ONIXDeviceDescriptor object cannot exist without first verifying that
a device of type ID exists at a specified address within the device
table
- This object is an argument to two public static methods ReadRegister
and WriteRegister within ONIDevice.cs (location is debatable)
- The idea is that if all register IO goes through these static methods,
then we are guaranteed to be talking to hardware that is actually there.
- One outstanding issues is that the I2CConfiguration class, and how it
is used, is pretty incompatible with this design.
- A second issue is that even though exceptions thrown by
ONIXDeviceDescriptor's constructor are generally caught and printed to
stderror, there is still a big UI performance hit. I'm not sure if there
is a way around this.